requirements.txt
amqp==2.6.1
asgiref==3.3.1
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.0
certifi==2020.12.5
chardet==4.0.0
chromedriver-install==1.0.3
coverage==5.1
Django==3.1.5
django-appconf==1.0.4
django-celery-beat==2.0.0
django-celery-email==3.0.0
django-environ==0.4.5
django-extensions==3.1.0
django-filter==2.4.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
EasyProcess==0.3
factory-boy==2.12.0
Faker==5.6.5
flower==0.9.5
gunicorn==20.0.4
humanize==3.2.0
idna==2.10
kombu==4.6.11
Markdown==3.3.3
more-itertools==8.6.0
packaging==20.8
pluggy==0.13.1
prometheus-client==0.8.0
psycopg2-binary==2.8.5
py==1.10.0
pyparsing==2.4.7
pytest==5.3.5
pytest-django==3.8.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
PyVirtualDisplay==2.0
redis==3.2.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
